+++ src/Kelly/Random/IRandomNumberGenerator.cs
@@ -1,6 +1,6 @@
 ﻿
 namespace Kelly.Random {
 	public interface IRandomNumberGenerator {
-		float Next();
+		double Next();
 	}
 }
